Add fuel tank and solid fuel support
No idea what "tank.level" was supposed to be, changed to capacity.

Always treat fuel as a tank, if we're solid-fueled then just pretend
it's a tank of size 64.

If a tank empties out and goes away, we still need to update it to
indicate that. This needs some testing.